Update App.tsx
TODO: It is incumbent upon me to transform this into a function, encapsulating all the keys I have laboriously generated for each domain I utilize  (This necessity arises, regrettably, from what appears to be a lack of understanding on the part of Adobe developers regarding the operation of CORS. To put it mildly, their management of authorizations leaves much to be desired. They seem to rely on a nonsensical one-to-one mapping system that, while it may be easier for the typical Adobe developer to program, lacks any logical coherence. This approach is not only perplexing but also deeply distressing).